+++ agentlib_mpc/models/serialized_ml_model/__init__.py
@@ -1,5 +1,4 @@
 # __init__.py
-from .base_serialized_model import MLModels, SerializedMLModel
-from .model_factory import get_model_class, register_model
+from .serialized_ml_model import MLModels, SerializedMLModel
 
-__all__ = ["MLModels", "SerializedMLModel", "get_model_class", "register_model"]
+__all__ = ["MLModels", "SerializedMLModel"]
